requirements.txt
uvicorn
requests
python-dotenv
sqlalchemy
mysql-connector-python
fastapi